+++ src/misc/buildInfo.ts
@@ -1,3 +1,3 @@
 export const REPO = "voces/fixus";
 export const VERSION = "12";
-export const BUILD = "2026-05-31T04:51:50.151Z";
+export const BUILD = "2026-05-31T06:10:45.737Z";